pyproject.toml
[tool.poetry]
name = "web-shop"
version = "0.1.0"
description = ""
authors = ["SunPower120 <112656928+SunPower120@users.noreply.github.com>"]
license = "MIT"
packages = [{include = "web_shop"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.4"
asgiref = "3.7.2"
certifi = "2023.7.22"
charset-normalizer = "3.2.0"
django-cors-headers = "4.2.0"
djangorestframework = "3.14.0"
django-rest-knox = "^4.2.0"
djoser = "^2.1.0"
idna = "3.4"
pytz = "2023.3"
pyyaml = "6.0.1"
requests = "2.31.0"
sqlparse = "0.4.4"
typing-extensions = "4.7.1"
tzdata = "2023.3"
urllib3 = "2.0.4"
psycopg2-binary = "^2.9.7"
drf-access-policy = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
